Separated OptionsPanel for Strategy execution and for backtest
