+++ src/components/BacktestOptionsPanel/index.js
@@ -1,0 +1,3 @@
+import BacktestOptionsPanel from './BacktestOptionsPanel'
+
+export default BacktestOptionsPanel
