+++ src/components/StrategyOptionsPanel/StrategyOptionsPanel.js
@@ -1,24 +1,20 @@
-import React, { useState } from 'react'
+import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
 import _find from 'lodash/find'
 import _size from 'lodash/size'
-import { Tooltip, Checkbox, Truncate } from '@ufx-ui/core'
+import { Tooltip } from '@ufx-ui/core'
 
 import _includes from 'lodash/includes'
 import { useTranslation } from 'react-i18next'
-import timeFrames from '../../util/time_frames'
-import DateInput from '../OrderForm/FieldComponents/input.date'
 
-import TimeFrameDropdown from '../TimeFrameDropdown'
-import AmountInput from '../OrderForm/FieldComponents/input.amount'
 import MarketSelect from '../MarketSelect'
 import Button from '../../ui/Button'
 import { makeShorterLongName } from '../../util/ui'
+import Dropdown from '../../ui/Dropdown'
 
 import './style.css'
 
 const MAX_STRATEGY_LABEL_LENGTH = 25
-const MAX_DATE = new Date()
 
 const StrategyOptionsPanel = ({
   strategy,
@@ -28,39 +24,32 @@ const StrategyOptionsPanel = ({
   markets,
   setMargin,
   setSymbol,
-  isPaperTrading,
-  timeframe,
-  setTimeframe,
-  trades,
-  setTrades,
-  margin,
-  candleSeed,
-  setCandleSeed,
   setFullScreenChart,
-  isBacktest,
-  onBacktestStart,
-  startDate,
-  endDate,
-  setStartDate,
-  setEndDate,
-  isBacktestLoading,
 }) => {
-  const [seedError, setSeedError] = useState(null)
-
   const { t } = useTranslation()
-
-  const updateSeed = (v) => {
-    const error = AmountInput.validateValue(v, t)
-    const processed = AmountInput.processValue(v)
-
-    setSeedError(error)
-    setCandleSeed(processed)
-  }
 
   const { label: strategyName } = strategy || {}
   const strategyDisplayName = strategyDirty
     ? t('strategyEditor.unsavedStartegy')
     : strategyName
+
+  const strategyTypesOptions = useMemo(() => {
+    return [
+      { value: 'momentum', label: t('strategyEditor.strategyTypes.momentum') },
+      {
+        value: 'meanReverting',
+        label: t('strategyEditor.strategyTypes.meanReverting'),
+      },
+      {
+        value: 'trendFollowing',
+        label: t('strategyEditor.strategyTypes.trendFollowing'),
+      },
+      {
+        value: 'technicalAnalysis',
+        label: t('strategyEditor.strategyTypes.technicalAnalysis'),
+      },
+    ]
+  }, [t])
 
   return (
     <div className='hfui-strategy-options'>
@@ -84,27 +73,6 @@ const StrategyOptionsPanel = ({
           )}
         </>
       </p>
-      {isBacktest && (
-        <>
-          <div className='hfui-strategy-options__input item'>
-            <DateInput
-              onChange={setStartDate}
-              def={{ label: t('strategyEditor.startDate') }}
-              value={startDate}
-              maxDate={endDate}
-            />
-          </div>
-          <div className='hfui-strategy-options__input item'>
-            <DateInput
-              onChange={setEndDate}
-              def={{ label: t('strategyEditor.endDate') }}
-              value={endDate}
-              maxDate={MAX_DATE}
-              minDate={startDate}
-            />
-          </div>
-        </>
-      )}
       <div className='hfui-strategy-options__input item'>
         <MarketSelect
           value={symbol}
@@ -118,70 +86,31 @@ const StrategyOptionsPanel = ({
           markets={markets}
           renderWithFavorites
         />
-        <p className='hfui-orderform__input-label'>{t('strategyEditor.selectMarketDescription')}</p>
+        <p className='hfui-orderform__input-label'>
+          {t('strategyEditor.selectMarketDescription')}
+        </p>
       </div>
       <div className='hfui-strategy-options__input item'>
-        <TimeFrameDropdown tf={timeframe} onChange={setTimeframe} />
-        <p className='hfui-orderform__input-label'>{t('strategyEditor.selectCandleDurationDescription')}</p>
+        <Dropdown options={strategyTypesOptions} />
+        <p className='hfui-orderform__input-label'>
+          {t('strategyEditor.strategyTypeDescription')}
+        </p>
       </div>
-      <AmountInput
-        className='hfui-strategy-options__amount-input item'
-        def={{ label: t('strategyEditor.candleSeedCount') }}
-        validationError={seedError}
-        value={candleSeed}
-        onChange={updateSeed}
-      />
-      {!isPaperTrading && _includes(symbol?.contexts, 'm') && (
-        <div className='hfui-strategy-options__amount-input item'>
-          <Checkbox
-            label={t('strategyEditor.useMarginCheckbox')}
-            checked={margin}
-            onChange={setMargin}
-          />
-          <p className='hfui-orderform__input-label'>{t('strategyEditor.useMarginCheckboxDescription')}</p>
-        </div>
-      )}
+
       <div className='hfui-strategy-options__amount-input item'>
-        <Checkbox
-          label={t('strategyEditor.useTradesCheckbox')}
-          checked={trades}
-          onChange={setTrades}
+        <Button
+          className='hfui-strategy-options__fullscreen-btn'
+          label={t('strategyEditor.fullscreenChartBtn')}
+          onClick={setFullScreenChart}
+          green
         />
-        <div className='hfui-orderform__input-label'>
-          <Truncate>
-            {t('strategyEditor.useTradesCheckboxDescription')}
-          </Truncate>
-        </div>
       </div>
-      {isBacktest ? (
-        <div className='hfui-strategy-options__amount-input item'>
-          <Button
-            className='hfui-strategy-options__fullscreen-btn'
-            label={t('ui.startBtn')}
-            onClick={onBacktestStart}
-            disabled={isBacktestLoading}
-            green
-          />
-        </div>
-      ) : (
-        <div className='hfui-strategy-options__amount-input item'>
-          <Button
-            className='hfui-strategy-options__fullscreen-btn'
-            label={t('strategyEditor.fullscreenChartBtn')}
-            onClick={setFullScreenChart}
-            green
-          />
-        </div>
-      )}
     </div>
   )
 }
 
 StrategyOptionsPanel.propTypes = {
   markets: PropTypes.objectOf(PropTypes.object).isRequired,
-  isPaperTrading: PropTypes.bool.isRequired,
-  timeframe: PropTypes.oneOf(timeFrames).isRequired,
-  setTimeframe: PropTypes.func.isRequired,
   symbol: PropTypes.objectOf(
     PropTypes.oneOfType([
       PropTypes.string,
@@ -191,27 +120,13 @@ StrategyOptionsPanel.propTypes = {
     ]),
   ).isRequired,
   setSymbol: PropTypes.func.isRequired,
-  trades: PropTypes.bool.isRequired,
-  setTrades: PropTypes.func.isRequired,
-  margin: PropTypes.bool.isRequired,
   setMargin: PropTypes.func.isRequired,
-  candleSeed: PropTypes.number.isRequired,
-  setCandleSeed: PropTypes.func.isRequired,
   onOpenSaveStrategyAsModal: PropTypes.func.isRequired,
   strategyDirty: PropTypes.bool.isRequired,
   strategy: PropTypes.shape({
     label: PropTypes.string,
   }).isRequired,
   setFullScreenChart: PropTypes.func.isRequired,
-  isBacktest: PropTypes.bool,
-  onBacktestStart: PropTypes.func,
-  isBacktestLoading: PropTypes.bool,
-}
-
-StrategyOptionsPanel.defaultProps = {
-  isBacktest: false,
-  isBacktestLoading: false,
-  onBacktestStart: () => { },
 }
 
 export default StrategyOptionsPanel
